feat: :sparkles: I created ProductContext to manage product data
I created a ProductContext component to manage product data throughout the application.
I set up a context using React's createContext and useEffect hooks, fetching product data from the API using Axios.
The data is stored in a state variable and provided to the rest of the app through the context provider.
The ProductContext is now available for consumption by child components.