feat: :sparkles: I wrapped Routes with ProductsProvider and created useProducts hook
I wrapped the Routes in the App component with the ProductsProvider to make product data accessible throughout the application.
Additionally, I created a custom hook named `useProducts` in the ProductsProvider component to simplify access to the product data from the context.
This hook allows child components to easily retrieve the products stored in the context.